+++ tallyInc/basicConstruct.h
@@ -3,7 +3,7 @@
  
  * File:   tallyInc/basicConstruct.h
  *
- * Copyright (c) 2004-2014 by Stuart Ansell
+ * Copyright (c) 2004-2015 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -60,8 +60,8 @@ class basicConstruct
   int checkItem(const mainSystem::inputParam&,
 		const size_t,const size_t,T&) const;
 
-  int getLinkIndex(const std::string&) const;
-  int getLinkIndex(const mainSystem::inputParam&,const size_t,
+  long int getLinkIndex(const std::string&) const;
+  long int  getLinkIndex(const mainSystem::inputParam&,const size_t,
 		   const size_t) const;
 
   std::vector<int> getCellSelection(const Simulation&,
